requirements.txt
selene==2.0.0rc8
pytest==8.1.1
appium-python-client==4.0.0
python-dotenv==0.20.0
allure-pytest==2.13.5